third_party/blink/web_tests/flag-specific/disable-layout-ng/svg/custom/use-nested-missing-target-removed-expected.txt

CONSOLE ERROR: Uncaught TypeError: Cannot set properties of null (setting 'id')
PASS
